attentionai-classroom/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
hypothesis
pytest
httpx
python-dotenv
